[TWINFACES-385] moved TiersTable to widgets/table/tiers
