+++ src/widgets/tables/tiers/index.ts
@@ -1,0 +1,1 @@
+export * from "./tiers";
